requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
dlib==19.24.2
imutils==0.5.4
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Werkzeug==2.3.7
gunicorn==21.2.0